[REFACTOR] Add endpoints to receive top10 data from sales implementation in payment microservice

+++ src/main/java/id/ac/ui/cs/advprog/hoomgroomcommerce/model/Product.java
@@ -17,7 +17,7 @@ public class Product {
     private String productName;
     private String productDescription;
     private String productImage;
-    private int productQuantity;
+    private Long productQuantity;
     private Double productPrice;
     private Double productDiscountPrice;
